requirements.txt
langchain
langchain_google_genai
streamlit
python-dotenv
streamlit==1.31.0
python-dotenv==1.0.0
google-generativeai==0.3.2
langchain==0.1.0
langchain-google-genai==0.0.5